refactor(162): move construction helpers into package
