+++ app/src/androidTest/java/org/cescfe/numpairs/feature/game/GameRouteTest.kt
@@ -16,12 +16,12 @@ import java.util.concurrent.atomic.AtomicReference
 import org.cescfe.numpairs.R
 import org.cescfe.numpairs.data.puzzle.seed.initialPuzzle
 import org.cescfe.numpairs.domain.puzzle.assignment.ResolvedOperandAssignment
+import org.cescfe.numpairs.domain.puzzle.construction.resolvedTile
 import org.cescfe.numpairs.domain.puzzle.model.Board
 import org.cescfe.numpairs.domain.puzzle.model.Operator
 import org.cescfe.numpairs.domain.puzzle.model.Puzzle
 import org.cescfe.numpairs.domain.puzzle.model.Strip
 import org.cescfe.numpairs.domain.puzzle.model.StripItem
-import org.cescfe.numpairs.domain.puzzle.resolvedTile
 import org.cescfe.numpairs.feature.game.presentation.GameUiState
 import org.cescfe.numpairs.feature.game.ui.GameScreenRobot
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
