+++ app/src/main/java/org/cescfe/numpairs/domain/puzzle/construction/PuzzleConstruction.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.domain.puzzle
+package org.cescfe.numpairs.domain.puzzle.construction
 
 import org.cescfe.numpairs.domain.puzzle.assignment.ResolvedOperandAssignment
 import org.cescfe.numpairs.domain.puzzle.model.Expression
